pref: 봉사 상세 페이지 공지사항 및 정기모임 기능향상
