+++ pages/volunteer/detail/index.js
@@ -1,0 +1,9 @@
+import VolunteerDetail from "../../../src/components/units/volunteer/detail/VolunteerDetail.container";
+
+export default function VolunteerDetailPage() {
+  return(
+    <>
+      <VolunteerDetail />
+    </>
+  )
+}
