--- pages/volunteer/recommend/detail/index.js
@@ -1,9 +1,0 @@
-import VolunteerDetail from "../../../../src/components/units/volunteer/recommend/detail/VolunteerDetail.container";
-
-export default function VolunteerDetailPage() {
-  return(
-    <>
-      <VolunteerDetail />
-    </>
-  )
-}
