+++ src/components/units/intro/Intro.styles.js
@@ -141,7 +141,7 @@ export const NextButtonBlock = styled.div`
 export const NextButtonItem = styled.button`
   background-color: #FF6636;
   border-radius: 30px;
-  font-size: 20px;
+  font-size: 24px;
   color: white;
   cursor: pointer;
 `;
