+++ src/components/units/signup/signup02/Signup02.styles.js
@@ -185,12 +185,19 @@ export const CheckEmailBtn = styled.button`
   cursor: pointer;
 `;
 
+export const InputVerificationCodeContainer = styled.div`
+  width: 344px;
+  height: 126px;
+  margin-top: 23px;
+  visibility: hidden;
+`;
+
 export const VerificationCodeTitle = styled.div`
   width: 100%;
   height: 21px;
   font-size: 18px;
   font-weight: 600;
-  margin-top: 24px;
+  
 `;
 
 export const VerificationCodeBlock = styled.div`
@@ -266,7 +273,8 @@ export const NextButtonBlock = styled.div`
 export const NextButtonItem = styled.button`
   background-color: #FF6636;
   border-radius: 30px;
-  font-size: 20px;
+  font-size: 24px;
+  font-weight: bold;
   color: white;
   cursor: pointer;
 `;
