--- src/components/units/volunteer/recommend/detail/VolunteerDetail.presenter.js
@@ -1,249 +1,0 @@
-import * as S from "./VolunteerDetail.styles";
-import Image from "next/image";
-
-export default function VolunteerDetailUI() {
-  return (
-    <>
-      <S.WrapperContents>
-        <S.VolunteerIntroContainer>
-          <S.IntroMainImgBlock>
-            <Image
-              src="/images/volunteer/volunteerDetail/volunteer_main_image.svg"
-              alt="volunteer_main_image"
-              width={343}
-              height={160}
-            />
-          </S.IntroMainImgBlock>
-          <S.IntroTitle>모임을 소개하는 한마디</S.IntroTitle>
-          <S.IntroDetail>해당 모임에 어울리는 문장을 간단하게 적어주세요</S.IntroDetail>
-        </S.VolunteerIntroContainer>
-        <S.AnnouncementContainer>
-          <S.AnnouncementTitleBlock>
-            <S.AnnouncementTitle>공지사항</S.AnnouncementTitle>
-            <S.RightArrowImgBlock>
-              <Image
-                src="/images/volunteer/volunteerDetail/right_arrow.svg"
-                alt="right_arrow"
-                width={44}
-                height={44}
-              />
-            </S.RightArrowImgBlock>
-          </S.AnnouncementTitleBlock>
-          <S.AnnouncementDetailContainer>
-            <S.AnnouncementDetailBlock style={{ backgroundColor: "#FEF8F2" }}>
-              <S.AnnouncementItems>
-                <S.CheckBox>
-                  <S.CheckImg>
-                    <Image
-                      src="/images/volunteer/volunteerDetail/check_icon.svg"
-                      alt="check_icon"
-                      width={22}
-                      height={15}
-                    />
-                  </S.CheckImg>
-                </S.CheckBox>
-                <S.AnnouncementDetail>
-                  이번주 봉사활동은 업체의 개인사정으로 인해 한 주<br />
-                  쉬어가니 착오 없으시길 바랍니다.<br />
-                </S.AnnouncementDetail>
-              </S.AnnouncementItems>
-            </S.AnnouncementDetailBlock>
-            <S.AnnouncementDetailBlock>
-              <S.AnnouncementItems>
-                <S.CheckBox>
-                  <S.CheckImg>
-                    <Image
-                      src="/images/volunteer/volunteerDetail/check_icon.svg"
-                      alt="check_icon"
-                      width={22}
-                      height={15}
-                    />
-                  </S.CheckImg>
-                </S.CheckBox>
-                <S.AnnouncementDetail>
-                  이번주 봉사활동은 업체의 개인사정으로 인해 한 주<br />
-                  쉬어가니 착오 없으시길 바랍니다.<br />
-                </S.AnnouncementDetail>
-              </S.AnnouncementItems>
-            </S.AnnouncementDetailBlock>
-            <S.AnnouncementDetailBlock>
-              <S.AnnouncementItems>
-                <S.CheckBox>
-                  <S.CheckImg>
-                    <Image
-                      src="/images/volunteer/volunteerDetail/check_icon.svg"
-                      alt="check_icon"
-                      width={22}
-                      height={15}
-                    />
-                  </S.CheckImg>
-                </S.CheckBox>
-                <S.AnnouncementDetail>
-                  이번주 봉사활동은 업체의 개인사정으로 인해 한 주<br />
-                  쉬어가니 착오 없으시길 바랍니다.<br />
-                </S.AnnouncementDetail>
-              </S.AnnouncementItems>
-            </S.AnnouncementDetailBlock>
-          </S.AnnouncementDetailContainer>
-        </S.AnnouncementContainer>
-        <S.MeetingContainer>
-          <S.MeetingTitleBlock>
-            <S.MeetingTitle>정기모임</S.MeetingTitle>
-            <S.RightArrowImgBlock>
-              <Image
-                src="/images/volunteer/volunteerDetail/right_arrow.svg"
-                alt="right_arrow"
-                width={44}
-                height={44}
-              />
-            </S.RightArrowImgBlock>
-          </S.MeetingTitleBlock>
-          <S.WrapperMeetingDetail>
-            <S.MeetingDetailContainer>
-              <S.MeetingDetailContentsBlock>
-                <S.DetailContentsDateBox>
-                  <S.DetailDate>2/12 (월)</S.DetailDate>
-                  <S.LeftDdayBox>
-                    D-<S.Day>1</S.Day>
-                  </S.LeftDdayBox>
-                </S.DetailContentsDateBox>
-                <S.MeetingName>유기견 구조활동</S.MeetingName>
-                <S.MeetingDetailInfoBox>
-                  <Image
-                    src="/images/volunteer/volunteerDetail/meeting_detail_main.svg"
-                    alt="meeting_detail_main"
-                    width={148}
-                    height={92}
-                  />
-                  <S.MeetingDetailInfo>
-                    <S.MeetingDetailItems>
-                      <S.DetailInfoName>일시</S.DetailInfoName>
-                      <S.DetailInfo>2/12 (월) 오전 11:00</S.DetailInfo>
-                    </S.MeetingDetailItems>
-                    <S.MeetingDetailItems>
-                      <S.DetailInfoName>위치</S.DetailInfoName>
-                      <S.DetailInfo>부산대역</S.DetailInfo>
-                    </S.MeetingDetailItems>
-                    <S.MeetingDetailItems>
-                      <S.DetailInfoName>비용</S.DetailInfoName>
-                      <S.DetailInfo>2천원(신규 인원제외)</S.DetailInfo>
-                    </S.MeetingDetailItems>
-                    <S.MeetingDetailItems>
-                      <S.DetailInfoName>인원</S.DetailInfoName>
-                      <S.DetailInfo>
-                        <S.ParticipatedPeople>12</S.ParticipatedPeople>
-                        /
-                        <S.MaximunPeople>12</S.MaximunPeople>
-                      </S.DetailInfo>
-                    </S.MeetingDetailItems>
-                  </S.MeetingDetailInfo>
-                </S.MeetingDetailInfoBox>
-                <S.UsersAndParticipateBox>
-                  <S.UsersItems>
-                    <S.UserImg>
-                      <Image
-                        src="/images/volunteer/volunteerDetail/user_1.svg"
-                        alt="user"
-                        width={25}
-                        height={30}
-                      />
-                    </S.UserImg>
-                    <S.UserImg style={{ right: "8px" }}>
-                      <Image
-                        src="/images/volunteer/volunteerDetail/user_2.svg"
-                        alt="user"
-                        width={25}
-                        height={30}
-                      />
-                    </S.UserImg>
-                    <S.UserImg style={{ right: "16px" }}>
-                      <Image
-                        src="/images/volunteer/volunteerDetail/user_3.svg"
-                        alt="user"
-                        width={30}
-                        height={30}
-                      />
-                    </S.UserImg>
-                  </S.UsersItems>
-                  <S.ParticipateBtn>참가하기</S.ParticipateBtn>
-                </S.UsersAndParticipateBox>
-              </S.MeetingDetailContentsBlock>
-            </S.MeetingDetailContainer>
-          </S.WrapperMeetingDetail>
-        </S.MeetingContainer>
-        <S.MemberContainer>
-          <S.MemberTitleBlock>
-            <S.MemberTitle>회원목록</S.MemberTitle>
-          </S.MemberTitleBlock>
-          <S.MemberListContainer>
-            <S.MemberBox>
-              <S.MemberIcon>
-                <Image
-                  src="/images/volunteer/volunteerDetail/member_icon.svg"
-                  alt="member_icon"
-                  width={49}
-                  height={49}
-                />
-              </S.MemberIcon>
-              <S.MemberName>김포포</S.MemberName>
-              <S.StatusBlock>
-                <S.Status>관리자</S.Status>
-              </S.StatusBlock>
-            </S.MemberBox>
-            <S.MemberBox>
-              <S.MemberIcon>
-                <Image
-                  src="/images/volunteer/volunteerDetail/member_icon.svg"
-                  alt="member_icon"
-                  width={49}
-                  height={49}
-                />
-              </S.MemberIcon>
-              <S.MemberName>가나다라마바사</S.MemberName>
-              <S.StatusBlock>
-                <S.Status style={{backgroundColor: "#240D05"}}>매니저</S.Status>
-              </S.StatusBlock>
-            </S.MemberBox>
-            <S.MemberBox>
-              <S.MemberIcon>
-                <Image
-                  src="/images/volunteer/volunteerDetail/member_icon.svg"
-                  alt="member_icon"
-                  width={49}
-                  height={49}
-                />
-              </S.MemberIcon>
-              <S.MemberName>프로 봉사러</S.MemberName>
-            </S.MemberBox>
-            <S.MemberBox>
-              <S.MemberIcon>
-                <Image
-                  src="/images/volunteer/volunteerDetail/member_icon.svg"
-                  alt="member_icon"
-                  width={49}
-                  height={49}
-                />
-              </S.MemberIcon>
-              <S.MemberName>김포포</S.MemberName>
-            </S.MemberBox>
-            <S.MemberBox>
-              <S.MemberIcon>
-                <Image
-                  src="/images/volunteer/volunteerDetail/member_icon.svg"
-                  alt="member_icon"
-                  width={49}
-                  height={49}
-                />
-              </S.MemberIcon>
-              <S.MemberName>김포포</S.MemberName>
-            </S.MemberBox>
-          </S.MemberListContainer>
-          <S.Blank />
-        </S.MemberContainer>
-        <S.NextButtonBlock>
-          <S.NextButtonItem>가입하기</S.NextButtonItem>
-        </S.NextButtonBlock>
-      </S.WrapperContents>
-    </>
-  )
-}
